pyproject.toml
[project]
name = "socketio-notify-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Anjali Georance",email = "anjaligeorance@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-socketio[asyncio] (>=5.13.0,<6.0.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "aiohttp (>=3.12.4,<4.0.0)",
    "aiosmtplib (>=4.0.1,<5.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
